+++ hostlib/hostLib/platform/linux/i2c_a7.c
@@ -33,19 +33,6 @@ static int default_axSmDevice_addr = 0x48;      // 7-bit address
 
 #define DEV_NAME_BUFFER_SIZE 64
 
-static int gBackoffDelay = 0;
-
-void resetBackoffDelay() {
-    gBackoffDelay = 0;
-}
-
-static void BackOffDelay_Wait() {
-    if (gBackoffDelay < 200 ) {
-        gBackoffDelay += 1;
-    }
-    usleep(gBackoffDelay * 1000);
-}
-
 /**
 * Opens the communication channel to I2C device
 */
@@ -93,8 +80,7 @@ i2c_error_t axI2CInit(void **conn_ctx, const char *pDevName)
         pdev_name = default_axSmDevice_name;
         dev_addr = default_axSmDevice_addr;
     }
-
-    LOG_D("I2CInit: opening %s\n", pdev_name);
+    LOG_I("I2CInit: opening %s:%x\n", pdev_name, dev_addr);
 
     if ((axSmDevice = open(pdev_name, O_RDWR)) < 0)
     {
@@ -391,19 +377,16 @@ i2c_error_t axI2CRead(void* conn_ctx, unsigned char bus, unsigned char addr, uns
     {
         //LOG_E("Failed Read data (nrRead=%d).\n", nrRead);
         rv = I2C_FAILED;
-        BackOffDelay_Wait();
     }
     else
     {
         if (nrRead == rxLen) // okay
         {
             rv = I2C_OK;
-            resetBackoffDelay();
         }
         else
         {
             rv = I2C_FAILED;
-            BackOffDelay_Wait();
         }
     }
     LOG_D("Done with rv = %02x ", rv);